cambios generales en el dash

+++ ws/v2/rutas.php
@@ -16,14 +16,16 @@ if (isset($_GET['action'])) {
         case 'delete':
             Delete($conn);
             break;
-
         case 'getByRepartidor':
             GetByRepartidor($conn);
             break;
         case 'getCustomView':
             getCustomView($conn);
             break;
-        default:
+        case 'getById':
+            getById($conn);
+            break;
+        default:    
         echo json_encode(array(
             'idEstatus' => -1,
             'data' => array(),
@@ -36,7 +38,7 @@ index($conn);
 
 function index($conn){
     //index
-    $sql = "SELECT * FROM tbl_ruteo";
+    $sql = "SELECT * FROM tbl_ruteo ORDER BY idRuteo DESC";
     $result = mysqli_query($conn, $sql);
     if (mysqli_num_rows($result) > 0) {
         $rutas = array();
@@ -59,8 +61,59 @@ function index($conn){
     mysqli_close($conn);
 }
 
+function getById($conn){
+    //get by id
+    $json = file_get_contents('php://input');
+    $obj = json_decode($json,true);
+    if($obj == null){
+        echo json_encode(array(
+            'idEstatus' => -1,
+            'data' => array(),
+            'mensaje' => 'Bad request.'
+        ));
+        return;
+    }
+
+    $idRuteo = $obj['idRuteo'];
+    $sql = "SELECT * FROM tbl_ruteo WHERE idRuteo = $idRuteo";
+    $result = mysqli_query($conn, $sql);
+    if (mysqli_num_rows($result) > 0) {
+        $rutas = array();
+        while ($row = mysqli_fetch_assoc($result)) {
+            $rutas[] = $row;
+        }
+        echo json_encode(array(
+            'idEstatus' => 1,
+            'data' => $rutas[0],
+            'mensaje' => 'OK'
+        ));
+    } else {
+        echo json_encode(array(
+            'idEstatus' => 0,
+            'data' => array(),
+            'mensaje' => 'No se encontraron rutas'
+        ));
+    }
+}
+
 function getCustomView($conn){
-    $sql = "SELECT * FROM rutas_view";
+    $querySearch = "";
+
+    $json = file_get_contents('php://input');
+    $obj = json_decode($json,true);
+    if($obj != null){
+        $query = $obj["query"];
+        $querySearch = " WHERE (Folio LIKE '%$query%' OR nbRepartidor LIKE '%$query%')";
+    }
+
+    $sql = "SELECT * FROM rutas_view ";
+
+    if($querySearch != ""){
+        $sql .= $querySearch;
+    }
+
+    $sql .= " ORDER BY idRuteo DESC";
+
     $result = mysqli_query($conn, $sql);
     if (mysqli_num_rows($result) > 0) {
         $rutas = array();
@@ -186,13 +239,32 @@ function GetByRepartidor($conn){
 }
 
 function Update($conn){
+    $json = file_get_contents('php://input');
+    $obj = json_decode($json,true);
+
+    if($obj == null){
+        echo json_encode(array(
+            'idEstatus' => -1,
+            'data' => array(),
+            'mensaje' => 'Bad request.'
+        ));
+        return;
+    }
+
     //update
-    $idRuteo = $_POST['idRuteo'];
-    $fecha = $_POST['fecha'];
-    $Folio = $_POST['Folio'];
-    $idRepartidor = $_POST['idRepartidor'];
-    $idEstatus = $_POST['idEstatus'];
-    $feFin = $_POST['feFin'];
+    $idRuteo = $obj['idRuteo'];
+    $fecha = $obj['fecha'];
+    $Folio = $obj['Folio'];
+    $idRepartidor = $obj['idRepartidor'];
+    $idEstatus = $obj['idEstatus'];
+    $feFin = null;
+
+    if($idEstatus == 7){
+        //7 = Finalizado
+        $feFin = date("Y-m-d H:i:s");
+    }else{
+        $feFin = null;
+    }
 
     //prevent sql injection
     $idRuteo = mysqli_real_escape_string($conn, $idRuteo);
@@ -202,7 +274,7 @@ function Update($conn){
     $idEstatus = mysqli_real_escape_string($conn, $idEstatus);
     $feFin = mysqli_real_escape_string($conn, $feFin);
 
-    $sql = "UPDATE tbl_ruteo SET fecha = '$fecha', Folio = '$Folio', idRepartidor = '$idRepartidor', idEstatus = '$idEstatus', feFin = '$feFin' WHERE idRuteo = '$idRuteo'";
+    $sql = "UPDATE tbl_ruteo SET fecha = '$fecha', Folio = '$Folio', idRepartidor = '$idRepartidor', idEstatus = '$idEstatus',feFin = '$feFin' WHERE idRuteo = '$idRuteo'";
 
     $result = mysqli_query($conn, $sql);
 
@@ -224,11 +296,47 @@ function Update($conn){
 }
 
 function Delete($conn){
+    $json = file_get_contents('php://input');
+    $obj = json_decode($json,true);
+
+    if($obj == null){
+        echo json_encode(array(
+            'idEstatus' => -1,
+            'data' => array(),
+            'mensaje' => 'Bad request.'
+        ));
+        return;
+    }
+
     //delete
-    $idRuteo = $_POST['idRuteo'];
+    $idRuteo = $obj['idRuteo'];
 
     //prevent sql injection
     $idRuteo = mysqli_real_escape_string($conn, $idRuteo);
+
+    //verificar si tiene paradas no se puede eliminar
+    $sql = "SELECT * FROM tbl_paradas WHERE idRuta = '$idRuteo'";
+    $result = mysqli_query($conn, $sql);
+    if (mysqli_num_rows($result) > 0) {
+        echo json_encode(array(
+            'idEstatus' => -1,
+            'data' => array(),
+            'mensaje' => 'No se puede eliminar la ruta porque tiene paradas.'
+        ));
+        return;
+    }
+
+    //verificar si el estatus es en proceso no se puede eliminar
+    $sql = "SELECT * FROM tbl_ruteo WHERE idRuteo = '$idRuteo' AND idEstatus = 5";
+    $result = mysqli_query($conn, $sql);
+    if (mysqli_num_rows($result) > 0) {
+        echo json_encode(array(
+            'idEstatus' => -1,
+            'data' => array(),
+            'mensaje' => 'No se puede eliminar la ruta porque esta en proceso.'
+        ));
+        return;
+    }
 
     $sql = "DELETE FROM tbl_ruteo WHERE idRuteo = '$idRuteo'";
